+++ CHANGELOG.md
@@ -38,6 +38,7 @@ The following changes have been completed, and will be included in the next tagg
     - added `Filepath.join()`
     - added `Filepath.parse()`
     - added `Filepath.relative()`
+    - added `Filepath.resolve()`
   - added `MakeFilepathOptions`
   - added `isFilepathData()`
   - added `makeFilepath()`